requirements.txt

Flask==2.3.3
python-telegram-bot==13.15
gunicorn==21.2.0
pandas==2.2.2
